+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 2.1.8 ##
 
 1.  IMGUI: Fix PropertyField not with `includeChildren: true` and lead to broken `ReferencePicker` 
+2.  `ValidateInput` now can receive the value and/or the index (if it's in a list/array) in the callback function.
 
 ## 2.1.7 ##